バリデーションの設定調整、単体コードの実装
・購入済み状態では削除・編集出来ない様にアクションの制限を実装
・バリデーションの設定調整
・購入済み・未購入状態におけるビューを実装
・Formオブジェクト(order_sending)に対する単体テストコードを実装

+++ app/controllers/items_controller.rb
@@ -23,7 +23,9 @@ class ItemsController < ApplicationController
   end
 
   def edit
-    redirect_to root_path unless current_user.id == @item.user_id
+    if current_user.id != @item.user_id || @item.order != nil
+      redirect_to root_path
+    end
   end
 
   def update
@@ -36,7 +38,7 @@ class ItemsController < ApplicationController
 
   def destroy
     item = Item.find(params[:id])
-    if current_user.id == item.user_id
+    if current_user.id == item.user_id && item.order == nil
       item.destroy
     end
     redirect_to root_path
